+++ packages/g-plugin-mobile-interaction/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @antv/g-plugin-mobile-interaction
 
+## 0.9.4
+
+### Patch Changes
+
+-   Updated dependencies [6757ccbd]
+    -   @antv/g-lite@1.2.4
+
 ## 0.9.3
 
 ### Patch Changes