+++ packages/g-plugin-zdog-canvas-renderer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-plugin-zdog-canvas-renderer
 
+## 1.2.4
+
+### Patch Changes
+
+-   Updated dependencies [6757ccbd]
+    -   @antv/g-lite@1.2.4
+    -   @antv/g-canvas@1.11.4
+
 ## 1.2.3
 
 ### Patch Changes